check that no connection Go routines are running after every test
